% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFuns.R
\name{buildOWDB}
\alias{buildOWDB}
\title{Build OW Database}
\usage{
buildOWDB(connection)
}
\arguments{
\item{connection}{A DBI connection object.}
}
\description{
This function uses the built-in scraping functions in the \code{owlstats} package to build a database of OWL Stats.
This function should be run once to build a new database. It should not be run every time you want to update your
database; for that, use \code{updateOWDB}.
}
\examples{
library(RMySQL)
mydb = dbConnect(MySQL(), user='user', password='password', dbname='owl', host='host')
buildOWDB(mydb)
}
\keyword{database}
\keyword{scraping}
