% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeroVals.R
\name{getHeroVals}
\alias{getHeroVals}
\title{Fetch Hero Values}
\usage{
getHeroVals(match.ids)
}
\arguments{
\item{match.ids}{A list of match.ids, like those returned by \code{getSchedule()}.}
}
\value{
A dataframe of cumulative statistics for the match.ids specified, split by match, map, and player.
}
\description{
This function grabs cumulative player stats, split by map, for matches included in match.ids.
These values are the total stats - like eliminations, deaths, and damage - accumulated over the course of a map and are not prorated to any time interval.
An ambigious time_played column is included but the relationship between the value from the API and the actual time of the map is murky.
Heroes played and stats per hero are available through the API but are not scraped by this function. There is no breakdown of time played on each hero, rendering the statistics included effectively useless.
}
\examples{
getHeroVals(21515)
getHeroVals(c(21485, 21494, 21521, 21419, 21514, 21446, 21515))
}
\keyword{scraping}
