% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchResults.R
\name{getMaps}
\alias{getMaps}
\title{Fetch Maps}
\usage{
getMaps()
}
\value{
A dataframe with details on every map tracked by the OWL API.
}
\description{
A helper function that pulls the details for all maps trakced by the OWL API.
}
\examples{
getMaps()
}
\keyword{helper}
\keyword{scraping}
