% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchResults.R
\name{getMatchResults}
\alias{getMatchResults}
\title{Fetch Match Results}
\usage{
getMatchResults(match.ids)
}
\arguments{
\item{match.ids}{A list of match.ids, like those fetched by \code{getSchedule()}.}
}
\value{
A dataframe representing the match results for all the match.ids passed in.
}
\description{
This function grabs the match results for a given list of match ids, and returns the information as a dataframe.
Included is match information, such as teams playing, match status, match score, and individual map scores and types.
Note that the number of possible maps included goes up to eight to ensure that no information is lost from playoff matches.
}
\examples{
getMatchResults(21515)
getMatchResults(c(21485, 21494, 21521, 21419, 21514, 21446, 21515))
}
\keyword{scraping}
