% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayers.R
\name{getPlayers}
\alias{getPlayers}
\title{Fetch Players}
\usage{
getPlayers(include_inactive = TRUE)
}
\value{
A dataframe of player information, such as id, name, and role.
}
\description{
This function grabs details about players. It can grab active and some inactive players if specified.
}
\examples{
getPlayers()
getPlayers(FALSE)
}
\keyword{scraping}
