% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerStats.R
\name{getPlayerStats}
\alias{getPlayerStats}
\title{Fetch Player Stats}
\usage{
getPlayerStats(season, postseason = FALSE)
}
\arguments{
\item{season}{SeasonID, usually the year of a given season, i.e. "2019".}

\item{postseason}{If true, returns postseason statistics instead of regular season statistics. Defaults to false.}
}
\value{
A dataframe representing the match results for all the match.ids passed in.
}
\description{
This function grabs player stats for a given season per ten minutes.
These values are across the entire season and are not tracked on a per-game basis, nor are they split by hero.
}
\examples{
getPlayerStats(2018)
getPlayerStats(2019,TRUE)
}
\keyword{scraping}
