% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSchedule.R
\name{getSchedule}
\alias{getSchedule}
\title{Fetch Schedule}
\usage{
getSchedule(startYear, endYear, team = "All")
}
\arguments{
\item{startYear}{The first year of the schedule you wish to scrape, like "2018".}

\item{endYear}{The last year of the schedule you wish to scrape, like "2019".}

\item{team}{If specified, returns the schedule for a given team, such as "NYE" or "SFS". If unspecified, returns the schedule for all teams in that time frame.}
}
\value{
A dataframe containing the schedule, match.ids, and results for the seasons and teams specified.
}
\description{
This function gets the OWL schedule for a given season(s), and can fetch the schedule for a given team.
Note that due to the limitations of the API, limiting your scrape to a given team will not improve the runtime of this function.
}
\examples{
getSchedule(2018,2018)
getSchedule(2018,2019,team="NYE")
}
\keyword{scraping}
