% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandings.R
\name{getStandings}
\alias{getStandings}
\title{Fetch Standings}
\usage{
getStandings(season, stage.no = NA)
}
\arguments{
\item{season}{SeasonID, usually the year of a given season, i.e. "2019".}

\item{stage.no}{The number of an OWL stage. Defaults to NA.}
}
\value{
A dataframe representing the standings of a given OWL season or stage. Includes match wins/losses and map wins/losses/ties.
}
\description{
This function pulls the standings for a given stage or seasons from the OWL API.
If no stage is specified, pulls for the given season. If a stage is specified, pulls standings for just that stage.
}
\examples{
getStandings(2018)
getStandings(2019,3)
}
\keyword{scraping}
