% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeams.R
\name{getTeams}
\alias{getTeams}
\title{Fetch Teams}
\usage{
getTeams()
}
\value{
A dataframe with information on OWL teams.
}
\description{
This function grabs a list of active teams tracked by the OWL API.
}
\examples{
getTeams()
}
\keyword{scraping}
