% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hero_vals}
\alias{hero_vals}
\title{Hero Vals}
\format{A data frame with hero variables for different players on different maps.}
\usage{
hero_vals
}
\description{
A dataframe of all player stats, grouped by map and hero. 
Includes hero specific stats, such as Rip-Tire Kills.
Data is wide-formatted.
}
\keyword{datasets}
