% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_map_stats}
\alias{match_map_stats}
\title{Match Map Stats}
\format{A data frame with team results for each map.}
\usage{
match_map_stats
}
\description{
A dataframe of the results of each map. Includes stats like completion percentage for control point maps.
}
\keyword{datasets}
