% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFuns.R
\name{updateOWDB}
\alias{updateOWDB}
\title{Update OW Database}
\usage{
updateOWDB(connection)
}
\arguments{
\item{connection}{A DBI connection object.}
}
\description{
This function uses the built-in scraping functions in the \code{owlstats} package to update a previously existing
OWL stats database.This function should not be run to build a new database; you should use \code{buildOWDB}.
This function overwrites all tables in the database except heroVals and matchResults; for those tables it appends
new values.
}
\examples{
library(RMySQL)
mydb = dbConnect(MySQL(), user='user', password='password', dbname='owl', host='host')
updateOWDB(mydb)
}
\keyword{database}
\keyword{scraping}
